Support for direct access mode (#2)
Enabled by passing `access="direct"` and `recl` keyword options to the FortranFile constructor. read and write operations on such files must include the `rec` keyword to select which record to read/write.

We assume that such files use no record markers at all, which is what gfortran and ifort seem to be doing. The ifort documentation mentions a "vms" option, which uses a single control byte at the start of each record, but this is currently not supported here.

Closes #1.